Initial work migrating federationreporting and federationcompliance into single plugin named reporting.
This represents the initial port and changes to packages etc. It does not represent execution of test cases which are probably still in error awaiting updates in federationcore (to become core) to remove nimble references and update packages.

Tests need to be executed when the above work is complete and further fine tuning executed.