Updated handling of determineDescriptorProtocolSupport.
This is now processed internally to the service instead of called by the controller to ensure that updates to the referenced descriptor occur before save. This in turn ensures that the descriptor is protocol valid before save in all circumstances.

Additionally this process is now called on endpoint update so that modification to endpoint protocol type is correctly reflected in the descriptor.