Dashboard: Adjusted reference to potentialApprovers.
Closes GH-30.

Previously the dashboard gsp specified a reference to task.potentialApprovers for iteration not allowing for the fact
that no approvers may be available if the task was created during some underlying error state. This change sets
evaluation to handle the case where no approvers are available with causing an exception.