branding: Corrects production fault relating to image in email layout.
Closes GH-31.

After several revisions using inbuilt grails resources links the email template
is still causing 500 errors in production. For some reason the underlying code
appends a / to the beginning of the complete URL causing it to throw an exception
 when passed to the Java URL class. This appears to be a known fault in Grails 2.0.3

This revision directly references the serverURL configured in fr-config.groovy to eliminate
reliance on resources subsystem and thus eliinate the error while still make the img src URL
dynamic to the deployed environment.