project: Refinement of commit policy based on initial feedback.
It was considered that the wording aroud commit message was to general and seemed to apply to atomic commits on individual feature branches.
This was not the intent so has been more explicitly stated to apply to patches/bugfixes being applied directly and all merge commits
directly on the 'develop' branch.