project: Improves migration between 1.x and 2.x.
Previously when migrating workflows the script generated did not force a flush of the initially deleted
workflow scripts. One some MySQL versions this was leading to an exception. Deletes are now force flushed.

Further to this by throwing an exception in the script instead of returning false vital clues where lost.
This has likewise been modified so that statements provided by println are presented in the console.