branding: Corrects rendering problems for IE8 users.
Closes #33.

IE8 is not correctly laying out the new FR 2 branding due to the low levels of support it has for some key standards.
As a result the modernizr library was re-introduced to all layouts (http://modernizr.com) to fix these issues in older browsers such as IE8.
UI has been validate to now render correctly in IE8 but without some nice extras such as rounded corners.