reporting: Correct ordering of IdP/SP selectables in compliance reports.
Closes #40.

The lists of IdP/SP presented within compliance reports are being ordered by ID and not displayName.
This is making it hard for end users to effectively report on the components they are interested in
as they need to look through the entire list.
The changes here modifies this to sort IdP/SP lists based on lowercase displayname.