branding: Organisation accepted email contained relative not absolute links.
Closes #47.

Fixes a problem in the shipping email templates for the final approval notification
for Organisations. Unfortunately links in the supplied email where relative
not absolute causing them to be useless for end users. An addition comma was
found to be the cause of this fault in the g:link tag syntax used.