app: Re-work loaded resources (css/js) for performance.
Closes #27.

Reworks loading of css/js to increase performance. Includes usage
of YUI minify to reduce size of output content. The number of unique
loads undertaken by the page is reduced by this commit however due to
what I consider a bug in the underlying resources plugin a jquery js
file is still loaded before css. I'll take this up with the resources
plugin developer with Grails to see if this can be resolved somehow.