foundation: Supplies sort order to Organisation Type values (primary and secondary).
Closes #49.

When displaying OrganisationType values associated with an Organisation (dropdowns and select boxes)
the ordering is now alphabetically based on displayName. In addition OrganisationTypes that are deemed
non useful for display in WAYF/DS interfaces are not presented allowing further disconnection of standalone
AA until they are properly supported in the codebase.