foundation: Corrects slow page load as a result of method used to handle admin addition for O, IdP, SP.
Closes #54.

The task of assigning administrative access within O, IdP, SP is based on standard form submission and does
not utilise ajax calls due to the infrequent requirement of this functionality and time limitations. Unfortunately
this approach requires a table with subjects known to FR to be populated at load time. As users increase this
does not scale adequately and slows down page load considerably, especially for IE. To solve this subjects are
now only loaded by ajax if an when an exisiting administrator decides to add more admins. This is still considered
an action that will be infrequently used. Standard form submission still applies from this point but may need
to be re-worked into a full ajax operation in the future.