foundation: Correctly associates transient NameID to new Identity Provider.
Closes #53.

On creation of a new IdP population of the default transient NameID was not
occuring. Selection of NameID has been removed from the default registration
wizard in order to streamline and simplify this process for new users but the
automatic association of the transient NameID had not yet been established in
the IdP service. In addition this change wraps a minor safe guard around default
NameID establishment in the equivalent SP service code.