app: Corrects a number of test cases and minor logic errors.
This commit solves a number of issues with test cases that where
as yet unfixed due to Grails migration and some internal FR changes (many relating to newer i18n space).

This commit does not however correct problems in WorkflowTaskServiceSpec which due to the way Grails 2 handles invocation of rendering the body of mail messages fails with "Cannot return Sitemesh factory it has not been set!". This is not a problem when actually running the application however as a proper request/response environment is active when workflow is creating all emails which is what is required.