Provide generic request and response verification/management.
Ensures that incoming requests for types POST, PUT, DELETE are
undertaken with a refer that is the local server to lower CSRF risks.

Ensures that all outbound requests are not cached as we run over SSL
and that OWASP recommended anti clickjaking headers are provided
as our app should never be embedded in a frame elsewhere.