Metadata: Extra checks required before rendering ACS for SP.
Previously an edge case existed for creating an SPSSODescriptor
and rendering an AttributeConsumingService as follows:
1. The SP had only 1 attribute or only 1 approved attribute.
2. That attribute required specification, commonly
eduPersonEntitlement whose approved values must be actually
specified in metadata.
3. Either the attribute had not specified the values it needed or
the values it was asking for was not yet approved.

In this scenario an 'empty' AttributeConsumingService was generated
breaking strict schema checks.

These commits correct this issue by ensuring RequestedAttribute
determines if it is functioning based on approval and checking it has
valid, approved values if specified. Metadata rendering then relies
on this assertion instead of making checks itself.

References #141.