Metadata: Switch WAYF should only see IdP who aren't marked as AA.
As part of the changes to introduce pure AA into our metadata this
DS configuration creation code was not updated.

This set of changes brings the configuration in line with metadata
so where we declare attributeAuthorityOnly as true the IdP is not
provided as a selectable option from the drop down list.

This should already have been part of 2.2.2 and was picked up in via
end user testing.

References #140.