Tests: Metadata: fix output from similarExcludingID
In metadata tests, fix the output from similarExcludingID.  If this
method call from a "then" clause of a Spock test returns false, the
output from the test (the error reported) includes only the call to this
method - result, method name, arguments.

Calling this method with a simple Diff object means only the the first
difference (which typically is the ID attribute ignored by this method)
is printed and the actual XML difference found is hidden away.

Change the code to instead convert the simple Diff to a DetailedDiff
before calling similarExcludingID - that way, all differences captured
in the DetailedDiff are printed and the test errors are much easier to
track down.