Tests: metadata: fix validUntil timezone to UTC
Some tests in the metadata plugin were using hard-coded timestamps but
these timestamps were interpreted relative to the current timezone - and
in other timezones then Australia/Brisbane caused the tests to fail as
the value would not match the UTC-pinned value in expected output.

Explicitly set the timezone to UTC in the tests and adjust test data
accordingly.