Metadata: RegistrationInfo: honour minimal and ext
Honour the minimal and ext (RoleExtensions) flags in the code generation
the Registration Info (RPI) extensions.  I.e., specifically, do not
generate this extension if either minimal is set OR roleExtensions is
not set.

Also pass these parameters to the populateSchema() method so that the
schema definition gets added ONLY if the RPI extension would be actually
generated.