Tests: AttributeFilter: fix output from similarExcludingID
Fix the similarExcludingID method in
AttributeFilterGenerationServiceSpec in a way similar to the fix in
856ebb4bd91210b6648477a0f9ed30843565003c.

A failed test would only print the first XML difference found which
would be the ID of the document - even though it's ignored for the
purpose of determining whether the test succeeds of fails.

Change the code to instead convert the simple Diff to a DetailedDiff
before calling similarExcludingID - that way, all differences captured
in the DetailedDiff are printed and the test errors are much easier to
track down.