AttributeFilter: support Regex in attribute values
Modify the AttributeFilterGenerationService to support regular
expressions in individual attribute values - relevant for
eduPersonEntitlement and highly needed for Groups.

Detect whether the value is a regular expression by checking whether it
starts with a "^" and ends with a "$" (same as what is done for Scope).

If the value is a regular expression, emit a basic:AttributeValueRegex
element instead of an basic:AttributeValueString element.

For groups, a suitable pattern is e.g.:

^urn:mace:groups.example.org:pip:association:example-association(?:|:.*)$

This follows the Java Regexp notation [1] and:
* releases the base value for the association
* all groups under that association (following a ":")
* does that as a non-capturing group.

[1] http://java.sun.com/j2se/1.5.0/docs/api/java/util/regex/Pattern.html

Solves #150
Solves #122