Ensure all export API requests are authenticated
All requests to the export controller are now run through a beforeInterceptor and authentication method. Clients must provide their identity and the configured shared secret in order to access data from the API.

In addition the export API functionality must be explicitly enabled, by default it is disabled.

New configuration options are documented in the example configuration file.