Messages: update tooltips about supporting regexp
Update tooltips as defined in messages.properties to inform about
support for regexp syntax in IdP Scope and eduPersonEntitlement values.

Code for these features has already been long in but the tooltips were
stating regexp syntax is not supported.

Also added a comment noting that the message
help.fr.serviceprovider.attribute.specvalue is unused
(and is also identical to help.fr.acs.specvalue)