Base: support for alert in IdP org selection
Add support for adding an alert to avoid multiple IdPs under a single
Organization.

For federations running a DS using the metadata as the source for
listing the IdPs (under their Organization name), add support to warn
users not to register multiple IdPs under the organization.

This is a (blue) alert box displaying above the Organization select
listbox on the IdP create page - but displaying only if the text is not
blank.

Default branding sets this to blank, but also has a commented out entry
with suitable text and instructions on when to uncomment the entry.