Branding: Welcome: Create Org/IdP/SP hint hooks
On the welcome page, add hooks above the Create Org/IdP/SP buttons to
allow putting additional hints in the branding.

In the default branding, set the hints to blank (no change), but also
provide commented out versions that provide suitable additional hints -
specifically targeting new users with instructions to FIRST register an
Organization, THEN proceed with registering an IdP or an SP.

No change to the UI if accepted with blank values included in the
default branding.