Metadata: support regexp scope also for AAuthority
In MetadataGenerationService, support regexp scope also for Attribute
Authority.

The initial support for regexp scope was introduced in
76c41461333334e13ac15125a84086f1d0b6a0f3 - but this was done only for
IDPSSODescriptorExtensions, not for
AttributeAuthorityDescriptorExtensions.

This omission has actually no impact, as in most cases (if
AttributeAuthorityDescriptor.collaborator is not null and points to an
IdPSSODescriptor), the collaborator descriptor is used as the
roleDescriptor for generating the AttributeAuthorityDescriptor and
therefore, IDPSSODescriptorExtensions gets used instead of
AttributeAuthorityDescriptorExtensions.

But fixing this to bring the code in sync.