Metadata: add basic support for MDUI
Add basic support for MDUI - add a mdui:UIInfo element to the Extensions
of a RoleDescriptor (IdPSSODescriptor, AttributeAuthorityDescriptor,
SPSSODescriptor) and populate it with mdui:DisplayName and
mdui:Description if the corresponding fields are set on the role
descriptor.

Do this only if explicitly enabled in the configuration.

Add new configuration entries to turn the rendering on or off
(separately for DisplayName and Description, defaulting to false for
both).  Add an additional configuration entry for setting the language.
(So far, the data model in FR allows only one language, set default
configuration entry to "en").

Add corresponding spec & test entries.

Solves #8

References:
* MDUI: http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-metadata-ui/v1.0/cs01/sstc-saml-metadata-ui-v1.0-cs01.html