Register ECP endpoint when creating an IdP
If the FR submission form presents ECP endpoint data in the params
object it will be created. ECP is becoming a default usecase for the
AAF and this process will ensure IdP are 'correct' from day one.

Apologies on whitespace correction by editor. Append ?w=1 on PR url to
hide from diff.