Export must mirror metadata collaborator decision
Previously the export functionality also looked at the flag
`attributeAuthorityOnly` on the collaborator instead of just the
collaborator detail directly when making the decision on how to
present exported data.

This differs from metadata generation which simply checks for the
presence of a collaborator and when found uses it to render all
relevant details, especially key data. MetadataGenerationService L364.

This commit synchronizes export to metadata, ensuring keys etc as they
currently appear within metadata will also appear in export JSON.