Upgrade Apache commons-collections 3.2.1 -> 3.2.2
Update FR to remove apache commons collections version that may be
vulnerable to "Mad Gadget"

https://opensource.googleblog.com/2017/03/operation-rosehub.html
https://www.kb.cert.org/vuls/id/576313

We believe that Grails was never directly impacted
http://stackoverflow.com/questions/41946971/grails-apache-commons-collection-vulnerability
but the most recent Google discussion does state "Mad Gadget is one of
the most pernicious vulnerabilities we’ve seen. By merely existing on the
Java classpath, seven “gadget” classes in Apache Commons Collections
(versions 3.0, 3.1, 3.2, 3.2.1, and 4.0) make object deserialization for
the entire JVM process Turing complete with an exec function."

Good enough for me and bu the extension the AAF to force this update.