Metadata: drop ds:KeyName from generated metadata
The trust model used in the construction of the metadata is based on
certificates explicitly included in metadata, not PKI-style trust
derived from trusted roots.

Key Name entered in Federation Registry is recorded only as a
descriptive name of the key / certificate.

Therefore, emitting the key name as ds:KeyName is superfluous here - and
more likely to cause troubles (if relying parties try to interpret it as
a name to be verified in a PKI-style chain from a trusted root).

Removing the ds:KeyName (both from shibmd:KeyAuthority and from KeyDescriptor).

Adjusting all tests accordingly (dropping ds:KeyName from expected output)

And fixing a typo in a test name.