Metadata: make AA roleDescriptor use AttributeAuthorityDescriptorExtensions
This is a follow-up to #235 - which had the intention to not emit MDUI
for an AttributeAuthorityDescriptor, but did not achieve that goal, as
the roleDescriptor method, even when rendering an AttributeAuthorityDescriptor,
was invoking the IDPSSODescriptorExtensions method to render the
Extensions, not AttributeAuthorityDescriptorExtensions.

This is because roleDescriptor constructs the name of the method to
invoke from the name of the class of the object passed in - and for an
AttributeAuthorityDescriptor, the collaborating IDPSSODescriptor gets
passed instead (when present), as that object has all the necessary
information (namely contacts and keyDescriptors).

This did not matter earlier - both Extensions method were identical.
The commit 81144d9 introduced in #235 changed
AttributeAuthorityDescriptorExtensions, but this method was not used.

To make roleDescriptor use AttributeAuthorityDescriptorExtensions,
I was considering passing the attributeAuthorityDescriptor object
instead - but this would not work due to the above (missing contacts and
keyDescriptors).

Further, I was hitting an issue with Hibernate where the object passed
was just the Hibernate proxy and NOT the real object.  This resulted
into the class name being seen as
"aaf.fr.foundation.AttributeAuthorityDescriptor_$$_javassist_79" and the
code constructing the Extensions method name broke.

There are (Hibernate-specific) ways to get around this, but again,
passing attributeAuthorityDescriptor directly would not work for the
above reasons.

Considering all of this, I have come up with the solutin to explicitly
pass the intended base class name as an additional parameter
("roleExtClass") to the roleDescriptor method.

This seems redundant for all other cases (IDPSSODescriptor,
SPSSODescriptor), but provides clean support for the use case where the
method gets an IDPSSODescriptor but should be treating it as an
AttributeAuthorityDescriptor.

It also remove the inspection of the class name, which might possibly
break with the above Hibernate issue also for other classes.

Successfully tested both with Grails tests and in deployment.