Allow AA to not have AttributeService endpoints
Strictly speaking this is a requirement in the SAML specification
however within FR AttributeAuthorityDescriptor is generally used in
collaboration with IDPSSODescriptors which now has valid reasons
for not exposing an AttributeAuthorityDescriptor.

This could already be achieved by disabling all AttributeService
endpoints but it is seen as useful to be able to actually delete all of
these endpoints to ensure a secondary administrator does not
accidentally re-enable them at some future time.