Allow PSE to be null to support 0 AS in AAD
The previous commit to enable AttributeAuthorityDescriptor to not supply
any AttributeService endpoints was broken because
protocolSupportEnumerations still expected to be present. Following
delete of the last AttributeService this will no longer be the case.

As per the last commit this is not to the letter of the law for the SAML
specification but it supports a necessary use case in FR for now.

Going forward tooling should support not providing an
AttributeAuthorityDescriptor in combination with an IDPSSODescriptor which
is 'correct' way to support this need.