base/Contact: show inactive Contacts in listings
Follow-up to #243 / c63e1d7 - show inactive contacts in main Contact
list and in Org/IdP/SP contact list.

This is two allow to:
1. find and re-active inactive contacts
2. actually remove association of inactive contacts with an entity

Note that inactive contacts are not shown in Search results when looking
for contacts to Add to an Org/IdP/SP.

And are not rendered in metadata.

This change is to allow cleaning up and/or reactivating contacts.