Views: disable autocomplete for invite code fields
Browser try with autocomplete by default - but this does not make sense
at all with the admin invite codes.

Browsers just unneccesarily store the past (no longer valid) values and
confusingly offer these to the user.

End this and hint the browsers with autocomplete="off".

Do this separately for the "descriptor" and "organization" templates.

And also mark the code field in the "organization" template as required
for consistency.