Grails2.5/Test/Metadata: avoid doubled-up SSOServices
IDPSSODescriptor defines the holder for singleSignOnServices as List:

    List singleSignOnServices

This makes singleSignOnServices behave as a list - and not a relationship.

Adding a SingleSignOnService element multiple times (via addToSingleSignOnServices)
makes it appear in the collection multiple times - and breaks things badly.

This has been present in the domain class behaviour already with Grails 2.2.4 - but
updating build-test-data to 2.4.0 just made it manifest in tests too.
(Presumably, build-test-data 2.4.0 now makes SingleSignOnService.build() adds the instance
to descriptor.singleSignOnService if a descriptor is provided).

In application code, addToSingleSignOnServices is called only in IdentityProviderService.create (3 times)
- and the SingleSignOnService instance is created without the descriptor being passed in.

Adjust tests to do the same: drop "descriptor:idp".
(Alternative was to comment out the idp.addToSingleSignOnServices(sso) call in these tests).