IDPService: avoid GORM executeUpdate in delete()
Deleting an IdP is tricky, because the IDPSSODescriptor object is cross-linked
with the AttributeAuthorityDescriptor object via mutual collaborator links.

The original solution - from commit 0e6492676 - uses a GORM executeUpdate call
to force the collaborator links to be changed to NULL and flushed to the database.

This commit replaced the original code which was just setting the links to null
- and which wasn't working because Hibernate optimization dropped the UPDATE
statements when it saw a DELETE coming for the same entity.

The GORM executeUpdate solution however does not work with MySQL servers running
in GTID mode, as GORM needs Temporary Tables for this kind of update,
and MySQL in GTID mode does NOT support Temporary Tables.

This commit switches back to the original solution (assigning collaborator
links to null in Groovy), and makes the following changes to make this work:

(1) add an explicit flush after updating the collaborator link
(2) WORKAROUND the fact that the flush only works for the
    AttributeAuthorityDescriptor but not the IDPSSODescriptor object,
    and reorder the deletes so that the single successful flush is
    sufficient to navigate the sequence of deletes correctly.