fix: base/forms: prevent forms with single input text from wrongly submitting
In forms with exactly one input of type text, browsers automatically submit the form when Enter is pressed.

However, for some of the forms (in particular related to endpoint manipulation),
this is a wrong thing to do.

The forms are not designed to be submitted (not in the literal sense).
They have a button with a Javascript handler that constructs an AJAX POST
message - but no action URL and method to submit.

The browser defaults to submitting to the location of the page (like SP details page)
via query string - which is just a very confusing user experience.
(The contents of the form disappears without being actually submitted anywhere).

Fix this by preventing these forms from being submitted with

    onsubmit="return false;"

Fixes REANNZ#7